Solution_Accelerators/Advanced_RAG/src/skills/ingestion/requirements.txt
python-dotenv==1.0.0
azure-identity==1.16.0
tenacity==8.4.1
pydantic==2.7.4
PyPDF2==3.0.1
shapely==2.0.3
Markdown==3.6
tabulate==0.9.0
pandas==2.0.3
numpy==1.26.4
openai==1.55.3
azure-storage-blob==12.19.1
azure-search-documents==11.4.0
azure-ai-documentintelligence==1.0.0b3
azure-keyvault==4.2.0
azure-keyvault-secrets==4.7.0
opencensus-ext-azure==1.1.9
opencensus==0.11.2
PyYAML==6.0.1
backoff==2.2.1
langchain-community==0.0.32
langchain==0.1.16
beautifulsoup4==4.12.3
aiohttp==3.9.1
aiohttp_cors==0.7.0
redis==5.0.4
setuptools==75.6.0